+++ src/compas_viewer/configurations/__init__.py
@@ -4,7 +4,7 @@ from .renderer_config import RendererConfig  # noqa: F401
 from .renderer_config import CameraConfig  # noqa: F401
 from .renderer_config import SelectorConfig  # noqa: F401
 from .controller_config import ActionConfig, MouseConfig  # noqa: F401
-from .scene_config import SceneConfig  # noqa: F401
+from .viewer_config import ViewerConfig  # noqa: F401
 from .layout_config import LayoutConfig  # noqa: F401
 from .layout_config import StatusbarConfig  # noqa: F401
 from .layout_config import WindowConfig  # noqa: F401