+++ src/compas_viewer/scene/brepobject.py
@@ -1,13 +1,18 @@
-from compas.geometry import Polyline
+from typing import Optional
+
+from compas.datastructures import Mesh
+from compas.geometry import Line
+from compas.geometry import Point
+from compas.scene import GeometryObject
+from compas.tolerance import TOL
 from compas.utilities import pairwise
 
-from .meshobject import MeshObject
-from .sceneobject import DataType
+from .geometryobject import GeometryObject as ViewerGeometryObject
 
 try:
     from compas_occ.brep import OCCBrep
 
-    class BRepObject(MeshObject):
+    class BRepObject(ViewerGeometryObject, GeometryObject):
         """Viewer scene object for displaying COMPAS OCCBrep geometry.
 
         Attributes
@@ -15,7 +20,7 @@ try:
         brep : :class:`compas_occ.brep.OCCBrep`
             The compas_occ Brep object.
         mesh : :class:`compas.datastructures.Mesh`
-            The tesselation mesh representation of the Brep.
+            The mesh representation of the Brep.
 
         See Also
         --------
@@ -23,38 +28,29 @@ try:
         """
 
         def __init__(self, brep: OCCBrep, **kwargs):
-            self.brep = brep
-            mesh, boundaries = self.brep.to_tesselation(kwargs.get("linear_deflection ", self.LINEARDEFLECTION))
-            super().__init__(mesh=mesh, **kwargs)
-            self.boundaries = boundaries
+            super().__init__(geometry=brep, **kwargs)
+            self.geometry: OCCBrep
+            self._viewmesh, self._boundaries = self.geometry.to_tesselation(TOL.lineardeflection)
 
-        def _read_lines_data(self) -> DataType:
-            positions = []
-            colors = []
-            elements = []
-            lines = []
-            for polyline in self.boundaries:
-                lines += pairwise(polyline.points)
-            count = 0
-            for i, (pt1, pt2) in enumerate(lines):
-                positions.append(pt1)
-                positions.append(pt2)
-                color = self.linescolor.get(i, self.linescolor["_default"])  # type: ignore
-                colors.append(color)
-                colors.append(color)
-                elements.append([count, count + 1])
-                count += 2
-            return positions, colors, elements
+        @property
+        def points(self) -> Optional[list[Point]]:
+            """The points to be shown in the viewer."""
+            return self.geometry.points
 
-        def to_viewmesh(self, linear_deflection=1):
-            """
-            Convert the OCCBrep to a view mesh.
-            """
+        @property
+        def lines(self) -> Optional[list[Line]]:
+            """The lines to be shown in the viewer."""
             lines = []
-            for edge in self.brep.edges:
-                if edge.is_line:
-                    lines.append(Polyline([edge.vertices[0].point, edge.vertices[-1].point]))
-            return self.brep.to_tesselation(linear_deflection=linear_deflection), lines
+            for polyline in self._boundaries:
+                for pair in pairwise(polyline.points):
+                    lines.append(Line(*pair))
+
+            return lines
+
+        @property
+        def viewmesh(self) -> Mesh:
+            """The mesh volume to be shown in the viewer."""
+            return self._viewmesh
 
 except ImportError:
     pass