+++ src/compas_viewer/scene/cylinderobject.py
@@ -1,12 +1,15 @@
-from math import pi
+from typing import Optional
 
 from compas.datastructures import Mesh
 from compas.geometry import Cylinder
+from compas.geometry import Line
+from compas.geometry import Point
+from compas.scene import GeometryObject
 
-from .meshobject import MeshObject
+from .geometryobject import GeometryObject as ViewerGeometryObject
 
 
-class CylinderObject(MeshObject):
+class CylinderObject(ViewerGeometryObject, GeometryObject):
     """Viewer scene object for displaying COMPAS Cylinder geometry.
 
     See Also
@@ -15,6 +18,19 @@ class CylinderObject(MeshObject):
     """
 
     def __init__(self, cylinder: Cylinder, **kwargs):
-        self.u = kwargs.get("u", int(2 * pi * cylinder.radius / self.LINEARDEFLECTION))
+        super().__init__(geometry=cylinder, **kwargs)
+        self.geometry: Cylinder
 
-        super().__init__(mesh=Mesh.from_shape(cylinder, u=self.u), **kwargs)
+    @property
+    def points(self) -> Optional[list[Point]]:
+        """The points to be shown in the viewer."""
+        return None
+
+    @property
+    def lines(self) -> Optional[list[Line]]:
+        return None
+
+    @property
+    def viewmesh(self):
+        """The mesh volume to be shown in the viewer."""
+        return Mesh.from_shape(self.geometry, u=self.u, triangulated=True)