+++ src/compas_viewer/scene/planeobject.py
@@ -1,13 +1,16 @@
+from typing import Optional
+
+from compas.datastructures import Mesh
 from compas.geometry import Frame
+from compas.geometry import Line
 from compas.geometry import Plane
 from compas.geometry import Point
 from compas.scene import GeometryObject
 
-from .sceneobject import DataType
-from .sceneobject import ViewerSceneObject
+from .geometryobject import GeometryObject as ViewerGeometryObject
 
 
-class PlaneObject(ViewerSceneObject, GeometryObject):
+class PlaneObject(ViewerGeometryObject, GeometryObject):
     """
     Viewer scene object for displaying COMPAS Plane geometry.
 
@@ -36,18 +39,21 @@ class PlaneObject(ViewerSceneObject, GeometryObject):
             Point(*self.frame.to_world_coordinates([-self.planesize, self.planesize, 0])),
         ]
 
-    def _read_lines_data(self) -> DataType:
-        positions = [
-            Point(*self.frame.to_world_coordinates([0, 0, 0])),
-            Point(*self.frame.to_world_coordinates([0, 0, self.planesize])),
+    @property
+    def points(self) -> Optional[list[Point]]:
+        """The points to be shown in the viewer."""
+        return None
+
+    @property
+    def lines(self) -> Optional[list[Line]]:
+        return [
+            Line(
+                Point(*self.frame.to_world_coordinates([0, 0, 0])),
+                Point(*self.frame.to_world_coordinates([0, 0, self.planesize])),
+            )
         ]
-        colors = [self.linescolor["_default"], self.linescolor["_default"]]
-        elements = [[0, 1]]
 
-        return positions, colors, elements
-
-    def _read_frontfaces_data(self) -> DataType:
-        return self.vertices, [self.facescolor["_default"]] * 4, [[0, 1, 2], [0, 2, 3]]
-
-    def _read_backfaces_data(self) -> DataType:
-        return self.vertices, [self.facescolor["_default"]] * 4, [[2, 1, 0], [3, 2, 0]]
+    @property
+    def viewmesh(self) -> Mesh:
+        """The mesh volume to be shown in the viewer."""
+        return Mesh.from_vertices_and_faces(self.vertices, [[0, 1, 2], [0, 2, 3]])