+++ src/compas_viewer/scene/polylineobject.py
@@ -1,12 +1,14 @@
+from typing import Optional
+
+from compas.geometry import Line
+from compas.geometry import Point
 from compas.geometry import Polyline
 from compas.scene import GeometryObject
-from compas.utilities import pairwise
 
-from .sceneobject import DataType
-from .sceneobject import ViewerSceneObject
+from .geometryobject import GeometryObject as ViewerGeometryObject
 
 
-class PolylineObject(ViewerSceneObject, GeometryObject):
+class PolylineObject(ViewerGeometryObject, GeometryObject):
     """Viewer scene object for displaying COMPAS Polyline geometry.
 
     See Also
@@ -17,30 +19,19 @@ class PolylineObject(ViewerSceneObject, GeometryObject):
     def __init__(self, polyline: Polyline, **kwargs):
         super().__init__(geometry=polyline, **kwargs)
         self.geometry: Polyline
+        self.show_lines = True
 
-    def _read_points_data(self) -> DataType:
-        positions = [point for point in self.geometry.points]
-        colors = [
-            self.pointscolor.get(i, self.pointscolor["_default"])  # type: ignore
-            for i, _ in enumerate(self.geometry.points)
-        ]
-        elements = [[i] for i in range(len(positions))]
-        return positions, colors, elements
+    @property
+    def points(self) -> Optional[list[Point]]:
+        """The points to be shown in the viewer."""
+        return self.geometry.points
 
-    def _read_lines_data(self) -> DataType:
-        positions = []
-        colors = []
-        elements = []
-        if self.geometry.is_closed:
-            lines = pairwise(self.geometry.points + [self.geometry.points[0]])
-        else:
-            lines = pairwise(self.geometry.points)
-        count = 0
-        for i, (pt1, pt2) in enumerate(lines):
-            positions.append(pt1)
-            positions.append(pt2)
-            colors.append(self.pointscolor.get(i, self.pointscolor["_default"]))  # type: ignore
-            colors.append(self.pointscolor.get(i, self.pointscolor["_default"]))  # type: ignore
-            elements.append([count, count + 1])
-            count += 2
-        return positions, colors, elements
+    @property
+    def lines(self) -> Optional[list[Line]]:
+        """The lines to be shown in the viewer."""
+        return self.geometry.lines
+
+    @property
+    def viewmesh(self):
+        """The mesh volume to be shown in the viewer."""
+        return None