+++ src/compas_viewer/scene/sphereobject.py
@@ -1,12 +1,15 @@
-from math import pi
+from typing import Optional
 
 from compas.datastructures import Mesh
+from compas.geometry import Line
+from compas.geometry import Point
 from compas.geometry import Sphere
+from compas.scene import GeometryObject
 
-from .meshobject import MeshObject
+from .geometryobject import GeometryObject as ViewerGeometryObject
 
 
-class SphereObject(MeshObject):
+class SphereObject(ViewerGeometryObject, GeometryObject):
     """Viewer scene object for displaying COMPAS Sphere geometry.
 
     See Also
@@ -15,6 +18,20 @@ class SphereObject(MeshObject):
     """
 
     def __init__(self, sphere: Sphere, **kwargs):
-        self.u = kwargs.get("u", int(2 * pi * sphere.radius / self.LINEARDEFLECTION))
-        self.v = kwargs.get("v", int(2 * pi * sphere.radius / self.LINEARDEFLECTION))
-        super().__init__(mesh=Mesh.from_shape(sphere, u=self.u, v=self.v), **kwargs)
+        super().__init__(geometry=sphere, **kwargs)
+        self.geometry: Sphere
+
+    @property
+    def points(self) -> Optional[list[Point]]:
+        """The points to be shown in the viewer."""
+        return [self.geometry.frame.point]
+
+    @property
+    def lines(self) -> Optional[list[Line]]:
+        """The lines to be shown in the viewer."""
+        return None
+
+    @property
+    def viewmesh(self) -> Mesh:
+        """The mesh volume to be shown in the viewer."""
+        return Mesh.from_shape(self.geometry, u=self.u, v=self.v, triangulated=True)