+++ src/compas_viewer/viewer.py
@@ -15,12 +15,12 @@ from compas_viewer.configurations import ActionConfig
 from compas_viewer.configurations import ControllerConfig
 from compas_viewer.configurations import LayoutConfig
 from compas_viewer.configurations import RendererConfig
-from compas_viewer.configurations import SceneConfig
+from compas_viewer.configurations import ViewerConfig
 from compas_viewer.controller import Controller
 from compas_viewer.layout import Layout
+from compas_viewer.scene.scene import ViewerScene
+from compas_viewer.scene.sceneobject import ViewerSceneObject
 from compas_viewer.utilities import Timer
-
-from .scene import ViewerScene as Scene
 
 
 class Viewer:
@@ -87,18 +87,15 @@ class Viewer:
         show_grid: Optional[bool] = None,
         configpath: Optional[str] = None,
     ):
-
-        self.started = False
-
         # Custom or default config
         if configpath is None:
             self.renderer_config = RendererConfig.from_default()
-            self.scene_config = SceneConfig.from_default()
+            self.viewer_config = ViewerConfig.from_default()
             self.controller_config = ControllerConfig.from_default()
             self.layout_config = LayoutConfig.from_default()
         else:
             self.renderer_config = RendererConfig.from_json(Path(configpath, "renderer.json"))
-            self.scene_config = SceneConfig.from_json(Path(configpath, "scene.json"))
+            self.viewer_config = ViewerConfig.from_json(Path(configpath, "scene.json"))
             self.controller_config = ControllerConfig.from_json(Path(configpath, "controller.json"))
             self.layout_config = LayoutConfig.from_json(Path(configpath, "layout.json"))
 
@@ -118,12 +115,16 @@ class Viewer:
         if show_grid is not None:
             self.renderer_config.show_grid = show_grid
 
+        # Viewer
+        self.config = self.viewer_config
+
         #  Application
+        self.started = False
         self.app = QCoreApplication.instance() or QApplication(sys.argv)
         self.window = QMainWindow()
 
         # Scene
-        self.scene = Scene(self, config=self.scene_config)
+        self.scene = ViewerScene(self, name=self.layout_config.window.title, context="Viewer")
 
         # Controller
         self.controller = Controller(self, self.controller_config)
@@ -139,15 +140,18 @@ class Viewer:
         self.timer: Timer
         self.frame_count: int = 0
 
+        #  Primitive
+        self.objects: list[ViewerSceneObject]
+
     # ==========================================================================
     # Scene
     # ==========================================================================
 
-    def add(self, item, **kwargs):
+    def add(self, **kwargs):
         """
         Add an item to the scene.
         This is a compatibility function for the old version of the viewer.
-        While :func:`compas_viewer.scene.ViewerScene.add` is the recommended way to add an item to the scene.
+        While :func:`compas.scene.Scene.add` is the recommended way to add an item to the scene.
 
         Returns
         -------
@@ -155,7 +159,7 @@ class Viewer:
             The scene object.
         """
 
-        return self.scene.add(item, **kwargs)
+        return self.scene.add(**kwargs)
 
     # ==========================================================================
     # Runtime